backend-logistica/requirements.txt
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.40.2
Flask==3.1.0
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gspread==6.1.4
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
playwright==1.49.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyee==12.0.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.42
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
